DTO objects-changed long to int in id class fields

+++ src/main/java/app/model/StudentSubjectDTO.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class StudentSubjectDTO {
     String subjectName;
-    long studentId;
+    int studentId;
 }
